chore: rename cmd/earthly to cmd/earth
