+++ .github/workflows/ci-scheduled-codeql-analysis.yml
@@ -79,7 +79,7 @@ jobs:
 
     # - run: |
     #     go version
-    #     go build ./cmd/earthly/*.go
+    #     go build ./cmd/earth/*.go
 
     - name: Perform CodeQL Analysis
       uses: github/codeql-action/analyze@bc0b696b4103f5fe60f15749af68a046868d511a # CodeQL Bundle v2.25.4
