+++ cmd/earth/app/create.go
@@ -3,9 +3,9 @@ package app
 import (
 	"fmt"
 
-	"github.com/EarthBuild/earthbuild/cmd/earthly/base"
-	"github.com/EarthBuild/earthbuild/cmd/earthly/common"
-	"github.com/EarthBuild/earthbuild/cmd/earthly/subcmd"
+	"github.com/EarthBuild/earthbuild/cmd/earth/base"
+	"github.com/EarthBuild/earthbuild/cmd/earth/common"
+	"github.com/EarthBuild/earthbuild/cmd/earth/subcmd"
 )
 
 // EarthApp encapsulates the core earth command-line application.
