+++ cmd/earth/common/common.go
