+++ cmd/earth/common/shared.go
