+++ cmd/earth/flag/global.go
@@ -6,7 +6,7 @@ import (
 	"time"
 
 	"github.com/EarthBuild/earthbuild/buildkitd"
-	"github.com/EarthBuild/earthbuild/cmd/earthly/common"
+	"github.com/EarthBuild/earthbuild/cmd/earth/common"
 	"github.com/EarthBuild/earthbuild/util/containerutil"
 	"github.com/urfave/cli/v3"
 )
