+++ cmd/earth/subcmd/doc_cmds.go
