+++ cmd/earth/subcmd/doc_cmds_test.go
